added custom hooks, fixed listing cards and fixed map not showing on load
