+++ src/components/VenueListing/VenueItem.tsx
@@ -16,15 +16,15 @@ const VenueItem = ( {venue, isSelected, onClick}: Props ) => {
 
   return (
     <>
-      <div className={`mb-4 flex cursor-pointer sm:min-w-[45%] min-w-[90%] flex-col gap-x-4 overflow-hidden rounded-lg border-b-2 p-4 shadow-md lg:flex-row ${isSelected ? "bg-primary/20" : "bg-slate-100"}`} onClick={handleClick}>
-        <figure className="h-auto w-full lg:max-w-[40%]">
+      <div className={`mb-4 flex cursor-pointer sm:min-w-[45%] min-w-[90%] flex-col gap-x-4 overflow-hidden rounded-lg border-b-2 p-4 shadow-sm xl:grid xl:grid-cols-2 ${isSelected ? "bg-primary/20 border-primary/20" : "bg-slate-100"}`} onClick={handleClick}>
+        <figure className="h-auto w-full">
           <img
             src={`/images/listing/${venue.id}.webp`}
             alt=""
-            className="aspect-auto h-full w-full rounded-lg"
+            className="aspect-auto h-full rounded-lg"
           />
         </figure>
-        <figcaption className="md:mt-1 mt-3 text-secondary flex flex-col grow lg:max-w-[55%] 2xl:max-w-fit md:grow-0">
+        <figcaption className="xl:mt-0 mt-5 text-secondary flex flex-col grow 2xl:max-w-fit md:grow-0 ">
           <h3 className="font-bold leading-none text-primary md:text-[3cqb] uppercase">
             {venue.city},{" "}
             <span className="font-light text-secondary md:text-[2cqb]">
@@ -38,7 +38,7 @@ const VenueItem = ( {venue, isSelected, onClick}: Props ) => {
             Mesualidades Desde
             <strong className="text-primary">  ${venue.price} MXN</strong>
           </p>
-          <div className="mt-4 flex w-full justify-end pr-4 text-primary md:text-[2cqb]">
+          <div className="mt-4 flex w-full justify-end pr-4 text-primary md:text-[2cqb] ">
             <Link
               to="/sedes/x"
               className="inline-flex items-center font-medium hover:underline"
