+++ src/hooks/useVenueSelection.ts
@@ -1,0 +1,28 @@
+import { useState, useEffect } from "react";
+import { useSearchParams } from "react-router-dom";
+import { venueListing } from "../Models/Venue";
+import Listing from "../consts/Listing";
+
+export function useVenueSelection() {
+  const [selectedVenue, setSelectedVenue] = useState<venueListing>({} as venueListing);
+  const [searchParams, setSearchParams] = useSearchParams();
+
+  const handleVenueItemClick = (cityID: string) => {
+    setSearchParams(`?sede=${cityID}`);
+  };
+
+  useEffect(() => {
+    const currentVenue = searchParams.get("sede");
+    if (currentVenue) {
+      const foundedVenue = Listing.find((venue) => venue.id === currentVenue);
+      if (foundedVenue) {
+        setSelectedVenue(foundedVenue);
+      }
+    }
+  }, [searchParams]);
+
+  return {
+    selectedVenue,
+    handleVenueItemClick
+  };
+}
